pull out const vec
